LPS-51081 Move OSDetector to "gradle-util"
