--- modules/sdk/gradle-plugins-maven-plugin-builder/src/com/liferay/gradle/plugins/maven/plugin/builder/util/OSDetector.java
@@ -1,41 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.gradle.plugins.maven.plugin.builder.util;
-
-import java.io.File;
-
-/**
- * @author Brian Wing Shun Chan
- */
-public class OSDetector {
-
-	public static boolean isWindows() {
-		if (_windows != null) {
-			return _windows.booleanValue();
-		}
-
-		if (File.pathSeparator.equals(";")) {
-			_windows = Boolean.TRUE;
-		}
-		else {
-			_windows = Boolean.FALSE;
-		}
-
-		return _windows.booleanValue();
-	}
-
-	private static Boolean _windows;
-
-}
