+++ source-formatter.properties
@@ -160,7 +160,7 @@
     run.outside.portal.excludes.paths=\
         /ant-bnd/,\
         /gradle-plugins-maven-plugin-builder/,\
-        /gradle-plugins-poshi-runner/,\
+        /gradle-util/,\
         /markdown-converter/,\
         /osgi-util/,\
         /portal-wab-extender/,\
